+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.2.0 (DEV)
+
+- Add: Optional serialization before writing to driver and unserialization after
+  reading from driver.
+- Add: Interface `WithPersistentOptions` is now generic with 2 optional type
+  parameters.
+
 ## 0.1.0 (2022-09-06)
 
 - First release.